refactor into index, tokens and widget files
