+++ packages/logconsole/src/tokens.ts
@@ -1,0 +1,110 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+import { Token } from '@phosphor/coreutils';
+
+import { ISignal } from '@phosphor/signaling';
+
+import { nbformat } from '@jupyterlab/coreutils';
+
+import { IRenderMimeRegistry } from '@jupyterlab/rendermime';
+
+import { IOutputAreaModel } from '@jupyterlab/outputarea';
+
+/* tslint:disable */
+/**
+ * The Logger Registry token.
+ */
+export const ILoggerRegistry = new Token<ILoggerRegistry>(
+  '@jupyterlab/logconsole:ILoggerRegistry'
+);
+
+export type ILoggerRegistryChange = 'append';
+
+/**
+ * A Logger Registry that registers and provides loggers by source.
+ */
+export interface ILoggerRegistry {
+  /**
+   * Get the logger for the specified source.
+   *
+   * @param source - The name of the log source.
+   *
+   * @returns The logger for the specified source.
+   */
+  getLogger(source: string): ILogger;
+  /**
+   * Get all loggers registered.
+   *
+   * @returns The array containing all registered loggers.
+   */
+  getLoggers(): ILogger[];
+
+  /**
+   * A signal emitted when the logger registry changes.
+   */
+  readonly registryChanged: ISignal<this, ILoggerRegistryChange>;
+}
+
+export interface ILogPayloadBase {
+  type: string;
+}
+
+export interface ITextLog extends ILogPayloadBase {
+  type: 'text';
+  data: string;
+}
+
+export interface IHtmlLog extends ILogPayloadBase {
+  type: 'html';
+  data: string;
+}
+
+export interface INotebookLog extends ILogPayloadBase {
+  type: 'notebook';
+  data: nbformat.IOutput;
+}
+
+export type ILogPayload = ITextLog | IHtmlLog | INotebookLog;
+
+export type ILoggerChange = 'append' | 'clear';
+
+/**
+ * A Logger that manages logs from a particular source.
+ */
+export interface ILogger {
+  /**
+   * Log an output to logger.
+   *
+   * @param log - The output to be logged.
+   */
+  log(log: ILogPayload): void;
+  /**
+   * Clear all outputs logged.
+   */
+  clear(): void;
+  /**
+   * Number of outputs logged.
+   */
+  readonly length: number;
+  /**
+   * Rendermime to use when rendering outputs logged.
+   */
+  rendermime: IRenderMimeRegistry;
+  /**
+   * A signal emitted when the log model changes.
+   */
+  readonly logChanged: ISignal<this, ILoggerChange>;
+  /**
+   * A signal emitted when the rendermime changes.
+   */
+  readonly rendermimeChanged: ISignal<this, void>;
+  /**
+   * The name of the log source.
+   */
+  readonly source: string;
+  /**
+   * Output Area Model used to manage log storage in memory.
+   */
+  readonly outputAreaModel: IOutputAreaModel;
+}
